Don’t run most pre-commit hooks on LICENSES/
Projects that comply with the REUSE Spec are supposed to put the text of
any legal tools that they use in the LICENSES/ folder [1]. Personally, I
just download the plain text version of whatever legal tool I’m using
and put it into the LICENSES/ folder. I try on purpose to never modify
them. If there was a problem with the text of one of the legal tools,
then that problem should be fixed upstream, not in my repo.

Additionally, I don’t necessarily have the legal permission to modify
legal tools. The GNU GPL, version 3, says “Everyone is permitted to copy
and distribute verbatim copies of this license document, but changing it
is not allowed” [2]. The Apache License, version 2.0, can be modified,
but you have to remove any references to “Apache” [3]. Similarly, 🅭🄍 can
be modified, but if you do modify it, you aren’t allowed to use any of
Creative Commons’s trademarks [4]. Sure, I could modify 🅭🄍, but I don’t
want to bother removing all of Creative Commons’s trademarks.

Finally, changes to legal tools should probably be reviewed by lawyers,
and I don’t want to pay to have that happen.

For all of those previous reasons, I don’t want to modify the files in
the LICENSES/ folder. This change makes it so that all pre-commit hooks
(other than the REUSE Tool) don’t get run on files in the LICENSES/
folder. It makes sense to run the REUSE Tool on files in the LICENSES/
folder since the REUSE Tool verifies REUSE compliance. Any other
pre-commit hooks might recommend changes to those files.

I added an exclude key for every pre-commit hook except for the REUSE
one. As far as I know, that is the most efficient way to accomplish what
I’m trying to accomplish [5].

[1]: <https://github.com/fsfe/reuse-docs/blob/0913b0a83b36c161966be1c5e70c81bdadfb8a69/spec.md#license-files>
[2]: <https://www.gnu.org/licenses/gpl-3.0.html#license-text>
[3]: <https://www.apache.org/foundation/license-faq.html#mod-license>
[4]: <https://wiki.creativecommons.org/wiki/Modifying_the_CC_licenses>
[5]: <https://stackoverflow.com/q/69106270/7593853>