Create EditorConfig
This will help me maintain a consistent style across all of my repos. In
a future commit, I’m going to enable a pre-commit hook that will
validate that files follow the EditorConfig’s rules.

I removed the note about indentation from the README. From this point
onwards, you’ll need to look at the EditorConfig file to figure out the
indentation width. This makes it so that the indentation width is both
human-readable and machine-readable.