Call email requirement a limit instead of a threshold
The word “limit” is more accurate. When I hear the word “threshold”, it
makes me think that something is going to happen after I surpass it.
Nothing really happens when you write an email that has more than
seventy-eight characters in one line. When I hear the word “limit”, it
makes me think that you shouldn’t surpass it. That interpretation is
more accurate to what the RFC actually says.

Additionally, the RFC itself calls this requirement a limit.