Add recommendation about line length
I chose the number seventy-two for a very specific reason. I accept
patches via git-send-email [1]. When you submit changes using
git-send-email, the diff for your commits is included in the body of one
or more emails [2].

According to the standard for the format of email messages [3]: “There
are two limits that this specification places on the number of
characters in a line.  Each line of characters MUST be no more than
998 characters, and SHOULD be no more than 78 characters, excluding the
CRLF.”

The number seventy-two was chosen to make it less likely that we go past
that seventy-eight character limit.

I could have just make seventy-eight the limit, but characters tend to
get put in front of lines. For example, git-diff will add a space, a
plus or a minus before every line [4]. People also sometimes quote diffs
which will add an additional “> ” at the beginning of the line.

[1]: <https://jasonyundt.website/submitting-patches#send>
[2]: <https://git-scm.com/book/en/v2/Distributed-Git-Contributing-to-a-Project#_project_over_email>
[3]: <https://www.rfc-editor.org/rfc/rfc5322.html#section-2.1.1>
[4]: <https://git-scm.com/docs/git-diff#_combined_diff_format>