Replace detect-secrets with gitleaks
In a future commit, I’m going to add something to
.pre-commit-config.yaml that looks like this:

    repo: https://github.com/Jayman2000/yamllint-pr
    rev: 8cedbee40606fcf9f9bd202206692e26dff854b7

detect-secrets was incorrectly detecting that hash as a secret, so I
tried allowlisting it [1]:

    repo: https://github.com/Jayman2000/yamllint-pr
    # pragma: allowlist nextline secret
    rev: 8cedbee40606fcf9f9bd202206692e26dff854b7

That should have worked, but it didn’t. At the moment, there are two
open issue about this problem [2][3].

I don’t have any particular preference for detect-secrets over any other
secret-detecting tool. I only chose it because it was the first
general-purpose secret-detecting tool on pre-commit’s list of supported
hooks [4]. I chose it as an experiment to see if I would like using it,
and it turns out that like I don’t like using it.

[1]: <https://github.com/Yelp/detect-secrets#inline-allowlisting>
[2]: <https://github.com/Yelp/detect-secrets/issues/435>
[3]: <https://github.com/Yelp/detect-secrets/issues/708>
[4]: <https://pre-commit.com/hooks.html>