Make repo-style-checker check for standard pre-commit hooks
I have multiple repos that contain Python code, and it’s difficult to
make sure that I enable all of the same hooks for each repository. Plus,
if I ever want to switch which hooks I use for Python, I have to
remember to change it in each of those repos.

By creating and enforcing a standard list of pre-commit hooks, I’m
decreasing the number of things that I have to remember.