Add detect-bad-unicode command
A future commit will turn this command into a pre-commit hook. It’s
unlikely that these code points would ever be used in a file for a good
purpose, so it makes sense to ban them. If they were to appear in a
file, it would probably be because of a mistake. Additionally,
detect-bad-unicode implicitly assumes that files use UTF-8 without a
BOM, so it indirectly works to help enforce the use of UTF-8 without a
BOM.