Also match .pyi files when looking for .py files
This will make sure that Python-related hints for contributors and
pre-commit hooks even if repos only contain stub files [1]. The main
motivation behind creating this commit is to prepare for a future commit
that will add Python stub files to this repo.

That future commit will add wcwidth [2] as a dependency. Unfortunately,
wcwidth doesn’t have any type hints yet [3], and there doesn’t seem to
be a PEP 561 [4] stub package for wcwidth. When a module isn’t typed and
there’s no stub package for that module, mypy’s recommendation is to
write your own stub files [5]. As a result, that future commit is going
to add .pyi files to this repo.

Since that future commit is going to add a new type of file to this
repo, I thought that it would make sense to preemptively add support for
that file type to repo-style-checker.

[1]: <https://mypy.readthedocs.io/en/stable/stubs.html>
[2]: <https://pypi.org/project/wcwidth/>
[3]: <https://github.com/jquast/wcwidth/pull/71>
[4]: <https://peps.python.org/pep-0561/>
[5]: <https://mypy.readthedocs.io/en/stable/running_mypy.html#missing-library-stubs-or-py-typed-marker>