Use an upstream version of pretty-format-toml
The only reason why I was using my own fork of
language-formatters-pre-commit-hooks was because this PR [1] hadn’t been
merged yet. Now that it’s been merged [2], there’s no reason for me to
use my fork of that project.

This commit makes an odd change to .pre-commit-config.yaml. The commit
makes it so that there’s two repos with the same URL, but different
revs. I could have combined those two repos, but then I would have to
use an unstable version of language-formatters-pre-commit-hooks for all
of its hooks. I only want to use an unstable version for the
pretty-format-toml hook.

[1]: <https://github.com/macisamuele/language-formatters-pre-commit-hooks/pull/208>
[2]: <https://github.com/macisamuele/language-formatters-pre-commit-hooks/pull/208#event-11497747354>