Make RSC allow no args on other hooks
Before this commit, repo-style-checker would fail in a very specific
situation. Imagine that jasons_pre_commit_hooks/__init__.py contained
this:

    PCR_EXAMPLE: Final = PreCommitRepoInfo(
        url='https://example.com/pre-commit-hooks.git',
        hook_ids=('foo',),
        args=('--example-argument',)
    )
    PRE_COMMIT_REPOS_BY_PATH: Final = (
        ('**', PCR_EXAMPLE),
    )

and that .pre-commit-config.yaml contained this:

    repos:
        -
            repo: https://example.com/pre-commit-hooks.git
            rev: v1.0.0
            hooks:
                -
                    id: foo
                    exclude: &default_excludes '^LICENSES/'
                    args: [--example-argument]
                -
                    id: bar
                    exclude: *default_excludes

Before this commit, repo-style-checker would give an error in that
situation because the bar hook doesn’t have an args key. It shouldn’t
give an error (foo already has the correct value for args, and we never
told it to check bar), but it did anyway. This commit fixes that
problem.

This commit was made in preparation for a future commit. I was running
into this problem while working on that future commit.