Change glob for pre-commit-update
This change shouldn’t have any user-facing effect. In fact, it doesn’t
even have any theoretical advantages because repo-style-checker already
requires a .pre-commit-config.yaml file. It’s more of a style thing. It
just makes more sense to say that pre-commit-update is required if there
are pre-commit hooks to be updated.