Read copying.md from a file
Before this commit, the copying.md template was a hard coded string
literal in jasons_pre_commit_hooks/__init__.py. This commit makes it so
that this repo’s copying.md file is read and then transformed so that is
can be used as a template.

The main goal of this commit is to make it easier to change copying.md.
As an added bonus, it also deduplicates some of the data in this repo
(we’re no longer storing an almost identical copy of copying.md inside
jasons_pre_commit_hooks/__init__.py).