Make RSC check .editorconfig’s contents
Once repo-style-checker is turned into a pre-commit hook, this commit
will make it so that all of my repos use the same EditorConfig file.
Using the same EditorConfig file across all of my repos will go a long
way towards making my code more consistent.

In order for it to be accessible to the jasons_pre_commit_hooks package,
.editorconfig was moved to jasons_pre_commit_hooks/editor_config.ini.
The name “editor_config.ini” was chosen for two reasons:

1. If it was named .editorconfig, then editors would erroneously assume
that it was the root EditorConfig file.
2. Ending the file name with .ini makes pretty-format-ini run on it
without any extra configuration.