Make FPTM’s module’s name consistent
Before this commit, the following rules were used to name most of the
modules in this repo:

1. Start with the command’s name (example: detect-bad-unicode).
2. Turn any hyphens into underscores (example: detect_bad_unicode).

This commit makes it so that the modules for all commands follow those
rules. The only command whose module didn’t follow those rules was
forbid-paths-that-match. When forbid-paths-that-match was first added, I
used “forbid_paths” for its identifier [1]. I had done that to avoid
adding a line to pyproject.toml that was longer than seventy-two
characters:

forbid-paths-that-match = "jasons_pre_commit_hooks:forbid_paths_that_match"

In hindsight, I probably should have just used forbid_paths_that_match
as an identifier to begin with. The main motivation behind the
seventy-two character limit is to avoid having git-send-email
inadvertently cause people to send emails that contain lines that are
longer than seventy-eight characters [2]. (The number seventy-eight
comes from the standard for the format of emails [3].) The line in
question would have only been seventy-five characters long. Even if you
were to include that line in a patch and then quote the patch, it would
still be less than seventy-eight characters long.

Additionally, a later commit [4] added a similar line that is
eighty-four characters long:

unreleased-commit-checker = "jasons_pre_commit_hooks.unreleased_commit_checker:main"

If unreleased-commit-checker is allowed to have a line that goes over
the limit, then why can’t forbid-paths-that-match do the same thing?

                                 ✱ ✱ ✱

Additionally, let’s take a look at the actual text of RFC 5322, the
standard that sets the seventy-eight character line length limit:

    “Each line of characters MUST be no more than 998 characters, and
    SHOULD be no more than 78 characters, excluding the CRLF.”

    Source: [3]

The word “SHOULD” here is defined by RFC 2119 [5]. According to RFC
2119,

    “SHOULD   This word, or the adjective "RECOMMENDED", mean that there
    may exist valid reasons in particular circumstances to ignore a
    particular item, but the full implications must be understood and
    carefully weighed before choosing a different course.”

    Source: [6]

OK, so let’s make sure that we understand the full implications of
ignoring the seventy-eight character limit for this line. RFC 5322 tells
us why the line length limit was set in the first place:

    “The more conservative 78 character recommendation is to accommodate
    the many implementations of user interfaces that display these
    messages which may truncate, or disastrously wrap, the display of
    more than 78 characters per line, in spite of the fact that such
    implementations are non-conformant to the intent of this
    specification (and that of [RFC5321] if they actually cause
    information to be lost).”

    Source: [3]

In other words, a bad user interface might make it difficult to read
patches that contain ones of the lines that this commit adds. That’s OK.
If a user really wants to read the line, then they can use git-am [7]
and git-show [8].

A related standard, RFC 6532, also comments on the seventy-eight
character line length limit:

    “The 78-character limit remains defined in terms of characters, not
    octets, since it is intended to address display width issues, not
    line-length issues.”

    Source: [9]

Again, if users have problems seeing messages because the line added by
this patch is too long, then they can just use git-am and git-show.
Making this codebase consistent and easy to understand is more important
than making sure that borderline non-conformant implementations don’t
break. (How many implementations actually have issues with lines longer
than seventy-eight characters anyway?)

                                 ✱ ✱ ✱

On an unrelated note, the commit also moves the
editorconfig-checker-disable line up so that it covers the entire
project.scripts table. It didn’t need to be moved that far up, but I did
so to make it look less awkward. Now, editorconfig-checker is disabled
at the beginning of one section and enabled at the start of the next
section.

[1]: 6c6148f (Add forbid-paths-that-match command, 2024-01-10)
[2]: .editorconfig
[3]: <https://www.rfc-editor.org/rfc/rfc5322.html#section-2.1.1>
[4]: 585aff2 (Create unreleased-commit-checker, 2024-01-27)
[5]: <https://www.rfc-editor.org/rfc/rfc5322.html#section-1.2.1>
[6]: <https://www.rfc-editor.org/rfc/rfc2119#section-3>
[7]: <https://git-scm.com/docs/git-am>
[8]: <https://git-scm.com/docs/git-show>
[9]: <https://www.rfc-editor.org/rfc/rfc6532.html#section-3.4>