Enable unreleased-commit-checker
The main motivation behind this commit is to make sure that I follow the
rules that I set out for myself in VERSIONING.md. Additionally, this
commit will help me prepare for a future commit. That future commit will
make repo-style-checker require that unreleased-commit-checker is
enabled if the repo contains a file named VERSIONING.md. This commit
makes it so that I’m already complying with that rule before I actually
enact it.