Always use stable version of JPCH
Before this change, this repo sometimes used pre-commit hooks from a
stable release of jasons-pre-commit-hooks and sometimes used pre-commit
hooks from an unstable version of jasons-pre-commit-hooks. There used to
be a good reason for this. When this repo first enabled the
unreleased-commit-checker hook [1], that hook was only available in
unstable versions of jasons-pre-commit hooks.

Yesterday, a new stable version of jasons-pre-commit-hooks was released.
That new stable version of jasons-pre-commit-hooks is the first stable
version that has the unreleased-commit-checker hook [2]. Now that
there’s a stable version that has all of the hooks that this repo uses,
there’s no need to use an unstable version at all.

[1]: f3046d6 (Enable unreleased-commit-checker, 2024-02-07)
[2]: <https://github.com/Jayman2000/jasons-pre-commit-hooks/releases/tag/v0.1.0>