Add --disable-hook option to repo-style-checker
I want to create a new repo, and I want to use repo-style-checker on
that repo. That new repo is going to be private, so it doesn’t make
sense for me to use the reuse pre-commit hook there. Before this change,
repo-style-checker required that all repos had the reuse hook enabled.
This change makes it so that I can create that new private repo without
having to use the reuse hook on it.