Don’t link to the original Markdown site
Before this change, repo-style-checker required that you included a Hint
for Contributors about using CommonMark for Markdown files. That Hint
for Contributors linked to both the CommonMark site and the original
Markdown site.

Linking to the original Markdown site was potentially confusing. The
original Markdown site describes a certain syntax for Markdown, but
people shouldn’t be looking at that description. Instead, they should be
looking at CommonMark site for a description of CommonMark’s syntax.
Additionally, if people are interested in the original description of
Markdown, then they can find that by going to the CommonMark site
anyway because the CommonMark site links to the Markdown site.

This commit removes the link to the original Markdown site.

Fixes #2.