Turn Unicode Standard quote into a block quote
Before this change, the quote from ISO/IEC 6429:1992 was formatted as a
block quote, but the quote from the Unicode Standard was formatted as an
inline quote. Both of them are about the same size, so it make sense for
both of them to be formatted the same. This commit turns the quote from
the Unicode Standard into a block quote so that it matches the quote
from ISO/IEC 6429:1992.