Move [7] footnote
This change makes the quote from ISO/IEC 6429:1992 more consistent with
the quote from the Unicode Standard. Before this change, the footnote
for the Unicode Standard quote was right after the word “says”, but the
footnote for the ISO/IEC 6429:1992 quote was in the previous sentence.
This change moves the footnote for the ISO/IEC 6429:1992 quote in order
to make it consistent with the Unicode Standard quote. As an added
bonus, this change also makes it more clear that there’s a citation for
the ISO/IEC 6429:1992 quote.