Replace nixpkgs-fmt citation with nixfmt citation
As of 2024-07-24, nixpkgs-fmt is no longer in development. It has been
replaced by nixfmt [1]. There’s no chance that I’m going to use
nixpkgs-fmt in the future, so mentioning it in that essay didn’t further
my point. This change fixes the essay by replacing the nixpkgs-fmt
citation with a nixfmt citation.

Fixes #4.

[1]: <https://github.com/nix-community/nixpkgs-fmt/commit/bdb15b4c7e0cb49ae091dd43113d0a938afae02c>