Fix order of link reference definitions
At the moment, when I run “pre-commit run --all”, the
unreleased-commit-checker hook gives me the following error:

    “There are 30 unreleased commits.
    The oldest unreleased commit is 1 months old.
    It’s time to do a release.”

In order to fix that error, I have to do a release. In order to do a
release, I have to add patch notes for the release to CHANGELOG.md. I
started working on a future commit that will add patch notes to
CHANGELOG.md, but while I was working on that commit I ran into a
problem.

At the moment, CHANGELOG.md contains three reference links. In the
actual document, the links appear in this order:

    - v0.1.0
    - v0.0.0
    - pre-version-0

Before this change, the link reference definitions for those links
appeared in this order:

    - v0.0.0
    - v0.1.0
    - pre-version-0

While I was working on the future commit that will add patch notes for
the version that I’m about to release, I needed to add another link
reference to CHANGELOG.md. It wasn’t clear to me where my new link
reference should be put. Should it be above v0.0.0? Should it be put
below v0.1.0?

This commit fixes that problem. It changes the order that the link
references are defined in. Now, the order of the link reference
definitions matches to order that the reference links appear in. When I
go to finish that future commit that adds patch notes for the new
version, I will know exactly where to add my new link reference
definition.