Update wording of CommonMark hint
461062f (Don’t link to the original Markdown site, 2024-08-06) updated
the version of the CommonMark hint that repo-style-checker expects, but
it didn’t update the version of that hint that’s in this repo’s README.
This change makes the version of the hint that’s in this repo’s README
match the version of the hint that repo-style-checker expects.

The main motivation behind this change is to make it easier to make a
future commit. That future commit will enable repo-style-checker for
this repo. Making this repo repo-style-checker–compliant now will make
it so that I don’t have to make this repo compliant in the future.