Create test for running pre-commit with dev shell
This repository’s flake’s dev shell is supposed to contain all of the
dependencies that are needed to run this repository’s pre-commit hooks.
Unfortunately, the dev shell doesn’t actually contain all of the
necessary dependencies.

This change adds a test that proves that the devShell doesn’t contain
all of the necessary dependencies. The main motivation behind this
change is to make it easier to create a future commit. That future
commit will make it so that the test succeeds. (At the moment, the test
fails).