Create stable branch
The motivation behind this change is to make it easier to use the latest
stable version of Jason’s Pre-commit Hooks as a flake input. I want to
be able to easily use stable versions of Jason’s Pre-commit Hooks as
flake inputs so that I can easily create flakes that have dev shells
that allow you to run repo-style-checker. Whenever I create a new
repository, I end up running repo-style-checker a bunch of times in
order to help me create the first batch of commits for the new
repository. I want to be able to easily create flakes that allow me to
run repo-style-checker so that I can create new repositories more
easily.

This change adds some text that’s inaccurate at the moment. It says:

> If you want to use this repository’s flake as an input for another
> flake, then I recommend using this repository’s stable branch.

At the moment, it doesn’t really make sense to use the stable branch as
a flake input because the stable branch doesn’t contain a flake.nix file
yet. That being said, I’m planning on doing a stable release that adds a
flake.nix file soon. After that new stable release is made, the text
that this change adds will be accurate.