Start using upstream yamllint pre-commit hook
The main motivation behind this change is to make it easier to create a
future commit. That future commit will update the version of Jason’s
Pre-commit Hooks that .pre-commit-config.yaml uses. The new version of
the repo-style-checker hook that’s in the new version of Jason’s
Pre-commit Hooks has a new requirement. The new requirement is that an
upstream version of the yamllint pre-commit hook is used.

Making this change now means that we won’t have to make it in the future
commit.