Stop using yamllint fork as a pre-commit hook
Before this change, this repository used two different versions of
yamllint. One was the upstream version of yamllint and the other was a
fork of yamllint. The only reason why a fork was being used was because
older versions of the repo-style-checker pre-commit hook required that a
fork of yamllint was used. Now that we’re no longer using an older
version of the repo-style-checker pre-commit hook, there’s no reason for
us to continue to use a fork of yamllint.