Make dev shell and its check use same stdenv
Before this change, the
checks.x86_64-linux.dev-shell-dependencies-for-pre-commit check should
have been failing, but it wasn’t. That check is supposed to make sure
that the corresponding dev shell contains all of the dependencies that
users need in order install this repository’s pre-commit hooks. At the
moment, the corresponding dev shell is missing at least one dependency
[1], so that check should fail, but before this change, it wasn’t
failing.

It wasn’t failing because the dev shell and the check were using
different stdenvs. This change makes sure that they both use the same
stdenv.

[1]: <https://github.com/Jayman2000/jasons-pre-commit-hooks/issues/42>