Remove nixfmt requirement
Before this change, repo-style-checker required that you used two
different hooks in order to format Nix files. Having two different
formatters is redundant, and it might cause conflicts. This change
removes one of the formatters so that there’s only one.

I chose to keep the “nix fmt” formatter instead of the “nixfmt” one. The
“nix fmt” one will be more pure (it will use a binary from a Nix
derivation).

Closes #45.