Create forbid-template-markers pre-commit hook
I want pretty much all of my Git repositories to use repo-style-checker
and (if they use Nix flakes) comply with Jason’s Nix Flake Style Guide
[1]. When I create a new repository, I have to do a bunch of work to get
the repository to the point where it passes all of repo-style-checker’s
tests and complies with Jason’s Nix Flake Style guide.

I want to reduce the amount of effort that it takes to create new
repositories. In order to do so, I’m going to create a template
repository. The template repository will pass all of
repo-style-checker’s tests (except for one, see below) and will be
Jason’s Nix Flake Style Guide compliant. Whenever I want to create a new
repository, I’ll just fork the template repository and replace the name
of the template repository with the name of the new project.

Here’s where we run into a problem. Off the top of my head, I can think
of two places where the project’s name should be written: README.md and
COPYING.md. In the future, there might end up being more places where
the project’s name is supposed to be written. I don’t want to create a
new project but forget to update the name of the project in one of those
places.

The solution to that problem: a new pre-commit hook named
forbid-template-markers. forbid-template-markers will give you an error
if any of the files in your project contain the string
“placeholder-х𭢸눳🕛𫨖-placeholder”. In README.md and COPYING.md, the
template project will be named “placeholder-х𭢸눳🕛𫨖-placeholder”.
repo-style-checker will eventually be updated so that it requires the
forbid-template-markers pre-commit hook. This means that if you fork the
template repository and forget to change its name somewhere, then you
will get a pre-commit error. This will (hopefully) help ensure that I
remember to update the name of the project everywhere.

This will also mean that one of the template repository’s pre-commit
hooks will fail even if you’re trying to modify the template repository
itself, but that’s probably fine.

                                 ✱ ✱ ✱

The name “placeholder-х𭢸눳🕛𫨖-placeholder” was chosen for two
reasons:

1. Including the word “placeholder” helps clue readers in to its
purpose.

2. Including five randomly chosen Unicode characters means that it’s
very unlikely that the string would appear somewhere else by
coincidence.

The string was randomly generated using this Nix flake:

    flake.nix:

        {
          inputs.nixpkgs.url = "github:NixOS/nixpkgs/nixos-25.05";
          outputs =
            { nixpkgs, self }:
            let
              system = "x86_64-linux";
              pkgs = nixpkgs.legacyPackages."${system}";
            in
            {
              formatter."${system}" = pkgs.nixfmt-tree;
              packages."${system}".default =
                pkgs.writers.writePython3Bin "random-unicode-string"
                  {
                  }
                  ''
                    import random
                    import unicodedata

                    # See help(chr).
                    MIN_UNICODE_CODE_POINT = 0
                    MAX_UNICODE_CODE_POINT = 0x10FFFF
                    # Characters in these categories wouldn’t work well for our
                    # purposes.
                    FORBIDDEN_CATEGORIES = frozenset({
                        "Mn",  # Nonspacing_Mark
                        "Zl",  # Line_Separator
                        "Zp",  # Paragraph_Separator
                        "Cc",  # Control
                        "Cf",  # Format
                        "Cs",  # Surrogate
                        "Co",  # Private_Use
                        "Cn"  # Unassigned
                    })

                    result = ""
                    for i in range(5):
                        while True:
                            code_point = random.randint(
                                MIN_UNICODE_CODE_POINT,
                                MAX_UNICODE_CODE_POINT
                            )
                            character = chr(code_point)
                            category = unicodedata.category(character)
                            if category not in FORBIDDEN_CATEGORIES:
                                break
                        print(
                          repr(character)
                          + f" U+{code_point:04X} "
                          + unicodedata.name(character, "no name")
                          + f". Category: {category}."
                        )
                        result += character
                    print("Final result:")
                    print(f"placeholder-{result}-placeholder")
                  '';
            };
        }

    flake.lock:

        {
          "nodes": {
            "nixpkgs": {
              "locked": {
                "lastModified": 1753489912,
                "narHash": "sha256-uDCFHeXdRIgJpYmtcUxGEsZ+hYlLPBhR83fdU+vbC1s=",
                "owner": "NixOS",
                "repo": "nixpkgs",
                "rev": "13e8d35b7d6028b7198f8186bc0347c6abaa2701",
                "type": "github"
              },
              "original": {
                "owner": "NixOS",
                "ref": "nixos-25.05",
                "repo": "nixpkgs",
                "type": "github"
              }
            },
            "root": {
              "inputs": {
                "nixpkgs": "nixpkgs"
              }
            }
          },
          "root": "root",
          "version": 7
        }

[1]: <https://codeberg.org/JasonYundt/jasons-nix-flake-style-guide>