Use two spaces for indentation in flake.lock
Before this change, I would always have to manually change flake.lock
after it was generated in order to make it use four spaces for
indentation. This change makes it so that I don’t have to manually
modify flake.lock after it’s generated. Hopefully, this will save me
time in the future.

Closes #28.