Enable ruff-checker pre-commit hook
The main motivation behind this change is to make it easier to create a
future commit. That future commit is going to update the version of
Jason’s Pre-commit Hooks that this repository uses. The version of
repo-style-check that comes with new version of Jason’s Pre-commit Hooks
will require that we use the ruff-checker pre-commit hook. Enabling the
ruff-checker pre-commit hook now means that we won’t have to enable it
later.