Add --ignore-path-pattern option to RSC
I recently tried adding a Git subtree [1] to a repository that uses
repo-style-checker. repo-style-checker considered the contents of the
subtree to be a part of the main repository which is technically true,
but it wasn’t helpful in this situation. repo-style-checker was telling
me that I needed to enable Python-related pre-commit hooks because the
subtree contained Python code. The main repository didn’t contain any
Python code so enabling Python-related pre-commit hooks would have been
the wrong thing to do here.

This new --ignore-path-pattern option will allow me to use Git subtrees
with repo-style-checker effectively. Now, I can just use
--ignore-path-pattern to get repo-style-checker to ignore the contents
of the subtree.

[1]: <https://man.archlinux.org/man/git-subtree.1>