Make RSC no longer require flake-lock-updater
For most of my Nix flakes, using both the flake-lock-updater and
nix-check-flake pre-commit hooks is redundant. repo-style-checker
already requires that the nix-check-flake pre-commit hook is run.
There’s no real benefit to requiring both.

See #47 for details. This commit is related to #47, but it doesn’t
completely deal with that issue.